Revert "healthd: charger: Add board overrides in mode_charger"
This reverts commit a186fba4a55fe730701f516ba940358313f8c3f9.

Since I will be building a recovery, I will use a twrp tree,
instead of the sonyxperiadev ones, which include the sony board
overrides for healthd
(https://github.com/sonyxperiadev/device-sony-common/tree/master/hardware/healthd),
and when they are not included, this reverted commit will lead
to a compilation failure due to undefined references.
Moreover, they will anyway be useless in my usecase, because
they aren't needed for a recovery.

Finally, as I am using this branch for the recovery building
purposes, revert the commit here.

Signed-off-by: Daniel Vasquez <danielgusvt@yahoo.com>